requirements.txt
configparser
pytesseract
Pillow
pyautogui
